Move all the definitions to their own files
