+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/AtomicMutationIndexTest.java
@@ -71,7 +71,7 @@ public class AtomicMutationIndexTest {
     @ParameterizedTest
     @MethodSource
     void indexScenariosTest(IndexScenario scenario, String indexType) throws Exception {
-        scenario.runTest(new AtomicMutationIndexDefinitionFactory(indexType),
+        scenario.runTest((groupingLength, syntheticType) -> new AtomicMutationIndexDefinition(indexType),
                 () -> {
                     FDBRecordContextConfig config = FDBRecordContextConfig.newBuilder()
                             .setTimer(new FDBStoreTimer())
