+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/PermutedMinMaxIndexTest.java
@@ -117,7 +117,7 @@ class PermutedMinMaxIndexTest extends FDBRecordStoreTestBase {
     @ParameterizedTest
     @MethodSource
     void indexScenariosTest(IndexScenario scenario, String indexType) throws Exception {
-        scenario.runTest(new PermutedMinMaxIndexDefinitionFactory(indexType),
+        scenario.runTest((groupingLength, syntheticType) -> new PermutedMinMaxIndexDefinition(indexType),
                 this::openContext,
                 FDBRecordStore.newBuilder()
                         .setKeySpacePath(path));
