+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/TextIndexTest.java
@@ -78,8 +78,6 @@ import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStore;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreTestBase;
 import com.apple.foundationdb.record.provider.foundationdb.FDBStoreTimer;
 import com.apple.foundationdb.record.provider.foundationdb.FDBStoredRecord;
-import com.apple.foundationdb.record.provider.foundationdb.IndexScanRange;
-import com.apple.foundationdb.record.provider.foundationdb.indexes.scenarios.IndexDefinition;
 import com.apple.foundationdb.record.provider.foundationdb.indexes.scenarios.IndexScenario;
 import com.apple.foundationdb.record.query.RecordQuery;
 import com.apple.foundationdb.record.query.expressions.AndOrComponent;
@@ -237,52 +235,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     @IndexScenarios
     void indexScenariosTest(IndexScenario scenario) throws Exception {
         scenario.runTest(
-                (groupingLength, syntheticType) -> new IndexDefinition() {
-                    private final String indexName = "textScenarioIndex";
-
-                    @Override
-                    public RecordMetaData getMetaData() {
-                        RecordMetaDataBuilder metaDataBuilder = RecordMetaData.newBuilder()
-                                .setRecords(TestRecordsTextProto.getDescriptor());
-                        metaDataBuilder.getRecordType(COMPLEX_DOC).setPrimaryKey(concatenateFields("group", "doc_id"));
-                        metaDataBuilder.addIndex(SIMPLE_DOC,
-                                new Index(indexName, field("text"), IndexTypes.TEXT));
-                        return metaDataBuilder.build();
-                    }
-
-                    @Override
-                    public List<Message> generateRecords(final int count) {
-                        return IntStream.range(0, count)
-                                .mapToObj(i -> (Message)SimpleDocument.newBuilder()
-                                        .setDocId(i)
-                                        .setGroup(i % 2)
-                                        .setText("term" + i)
-                                        .build())
-                                .collect(Collectors.toList());
-                    }
-
-                    @Override
-                    public RecordCursor<IndexEntry> scanIndex(final FDBRecordStore store, final ScanProperties scanProperties) {
-                        return store.scanIndex(store.getRecordMetaData().getIndex(indexName),
-                                new IndexScanRange(BY_TEXT_TOKEN, TupleRange.ALL), null, scanProperties);
-                    }
-
-                    @Override
-                    public List<Message> generateOtherRecords(final int count) {
-                        // A document with no text produces no text-index tokens, so it is not covered by the index.
-                        return IntStream.range(0, count)
-                                .mapToObj(i -> (Message)SimpleDocument.newBuilder()
-                                        .setDocId(1000 + i)
-                                        .setGroup(i % 2)
-                                        .build())
-                                .collect(Collectors.toList());
-                    }
-
-                    @Override
-                    public String getIndexName() {
-                        return indexName;
-                    }
-                },
+                (groupingLength, syntheticType) -> new TextIndexDefinition(),
                 this::openContext,
                 FDBRecordStore.newBuilder()
                         .setKeySpacePath(path));
