+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/VersionIndexTest.java
@@ -70,7 +70,6 @@ import com.apple.foundationdb.record.provider.foundationdb.IndexScanBounds;
 import com.apple.foundationdb.record.provider.foundationdb.IndexScanRange;
 import com.apple.foundationdb.record.provider.foundationdb.KeyValueCursor;
 import com.apple.foundationdb.record.provider.foundationdb.ScanNonReadableIndexException;
-import com.apple.foundationdb.record.provider.foundationdb.indexes.scenarios.IndexDefinition;
 import com.apple.foundationdb.record.provider.foundationdb.indexes.scenarios.IndexScenario;
 import com.apple.foundationdb.record.provider.foundationdb.keyspace.KeySpacePath;
 import com.apple.foundationdb.record.query.RecordQuery;
@@ -486,49 +485,7 @@ public class VersionIndexTest {
     @IndexScenarios
     void  indexScenariosTest(IndexScenario scenario) throws Exception {
         scenario.runTest(
-                (groupingLength, syntheticType) -> new IndexDefinition() {
-
-                    private String indexName = "MySimpleRecord$num2-version";
-
-                    @Override
-                    public RecordMetaData getMetaData() {
-                        RecordMetaDataBuilder metaDataBuilder = RecordMetaData.newBuilder()
-                                .setRecords(TestRecords1Proto.getDescriptor());
-                        metaDataBuilder.setStoreRecordVersions(true);
-                        metaDataBuilder.addIndex("MySimpleRecord", new Index(indexName, VersionKeyExpression.VERSION, IndexTypes.VERSION));
-                        return metaDataBuilder.build();
-                    }
-
-                    @Override
-                    public List<Message> generateRecords(final int count) {
-                        return IntStream.range(0, count)
-                                .mapToObj(i -> (Message) TestRecords1Proto.MySimpleRecord.newBuilder()
-                                        .setRecNo(i).setNumValue2(-1 * i)
-                                        .build())
-                                .toList();
-                    }
-
-                    @Override
-                    public List<Message> generateOtherRecords(final int count) {
-                        // MyOtherRecord is not covered by the version index on MySimpleRecord.
-                        return IntStream.range(0, count)
-                                .mapToObj(i -> (Message) TestRecords1Proto.MyOtherRecord.newBuilder()
-                                        .setRecNo(1000 + i)
-                                        .build())
-                                .toList();
-                    }
-
-                    @Override
-                    public RecordCursor<IndexEntry> scanIndex(final FDBRecordStore store, ScanProperties scanProperties) {
-                        return store.scanIndex(store.getRecordMetaData().getIndex(indexName),
-                                new IndexScanRange(IndexScanType.BY_VALUE, TupleRange.ALL), null, scanProperties);
-                    }
-
-                    @Override
-                    public String getIndexName() {
-                        return indexName;
-                    }
-                },
+                (groupingLength, syntheticType) -> new VersionIndexDefinition(),
                 () -> {
                     FDBRecordContextConfig config = FDBRecordContextConfig.newBuilder()
                             .setTimer(new FDBStoreTimer())
