+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/leaderboard/LeaderboardIndexTest.java
@@ -24,7 +24,6 @@ import com.apple.foundationdb.record.EndpointType;
 import com.apple.foundationdb.record.EvaluationContext;
 import com.apple.foundationdb.record.ExecuteProperties;
 import com.apple.foundationdb.record.FunctionNames;
-import com.apple.foundationdb.record.IndexEntry;
 import com.apple.foundationdb.record.IndexScanType;
 import com.apple.foundationdb.record.IsolationLevel;
 import com.apple.foundationdb.record.RecordCursor;
@@ -53,7 +52,6 @@ import com.apple.foundationdb.record.provider.foundationdb.FDBStoredRecord;
 import com.apple.foundationdb.record.provider.foundationdb.IndexOrphanBehavior;
 import com.apple.foundationdb.record.provider.foundationdb.indexes.IndexScenarios;
 import com.apple.foundationdb.record.provider.foundationdb.indexes.RankedSetHashFunctions;
-import com.apple.foundationdb.record.provider.foundationdb.indexes.scenarios.IndexDefinition;
 import com.apple.foundationdb.record.provider.foundationdb.indexes.scenarios.IndexScenario;
 import com.apple.foundationdb.record.provider.foundationdb.keyspace.KeySpacePath;
 import com.apple.foundationdb.record.query.RecordQuery;
@@ -139,69 +137,7 @@ public class LeaderboardIndexTest {
     @IndexScenarios
     void indexScenariosTest(IndexScenario scenario) throws Exception {
         scenario.runTest(
-                (groupingLength, syntheticType) -> new IndexDefinition() {
-                    private final String indexName = "LeaderboardIndex";
-
-                    @Override
-                    public RecordMetaData getMetaData() {
-                        RecordMetaDataBuilder metaDataBuilder = RecordMetaData.newBuilder()
-                                .setRecords(TestRecordsLeaderboardProto.getDescriptor());
-                        metaDataBuilder.addIndex("NestedLeaderboardRecord",
-                                new Index(indexName,
-                                        Key.Expressions.field("scores", KeyExpression.FanType.FanOut)
-                                                .nest(Key.Expressions.concat(Key.Expressions.field("score"),
-                                                        Key.Expressions.field("timestamp")))
-                                                .ungrouped(),
-                                        IndexTypes.TIME_WINDOW_LEADERBOARD));
-                        return metaDataBuilder.build();
-                    }
-
-                    @Override
-                    public void setupIndex(final FDBRecordStore store) {
-                        // A leaderboard index needs at least one leaderboard (window) to exist before records
-                        // can be indexed. Create an all-time leaderboard (no explicit time-window specs).
-                        store.performIndexOperation(indexName, new TimeWindowLeaderboardWindowUpdate(
-                                System.currentTimeMillis(), true, 0L, true, Collections.emptyList(),
-                                TimeWindowLeaderboardWindowUpdate.Rebuild.NEVER));
-                    }
-
-                    @Override
-                    public List<Message> generateRecords(final int count) {
-                        return IntStream.range(0, count)
-                                .mapToObj(i -> (Message)TestRecordsLeaderboardProto.NestedLeaderboardRecord.newBuilder()
-                                        .setName("player-" + i)
-                                        .setGameId("game-1")
-                                        .addScores(TestRecordsLeaderboardProto.NestedLeaderboardEntry.newBuilder()
-                                                .setScore(100L + i)
-                                                .setTimestamp(1000L + i)
-                                                .setContext(i))
-                                        .build())
-                                .collect(Collectors.toList());
-                    }
-
-                    @Override
-                    public RecordCursor<IndexEntry> scanIndex(final FDBRecordStore store, final ScanProperties scanProperties) {
-                        return store.scanIndex(store.getRecordMetaData().getIndex(indexName),
-                                new TimeWindowScanRange(TimeWindowLeaderboard.ALL_TIME_LEADERBOARD_TYPE, 0L, TupleRange.ALL),
-                                null, scanProperties);
-                    }
-
-                    @Override
-                    public List<Message> generateOtherRecords(final int count) {
-                        // FlatLeaderboardRecord is not covered by the leaderboard index on NestedLeaderboardRecord.
-                        return IntStream.range(0, count)
-                                .mapToObj(i -> (Message)TestRecordsLeaderboardProto.FlatLeaderboardRecord.newBuilder()
-                                        .setName("other-" + i)
-                                        .setGameId("game-1")
-                                        .build())
-                                .collect(Collectors.toList());
-                    }
-
-                    @Override
-                    public String getIndexName() {
-                        return indexName;
-                    }
-                },
+                (groupingLength, syntheticType) -> new LeaderboardIndexDefinition(),
                 this::openContext,
                 FDBRecordStore.newBuilder()
                         .setKeySpacePath(path));
